ssztypes, codegen: measure a bit count by its bytes, widen every runtime field index
A length in bits was compared with the byte-domain size limit, so a vector of
2^32 bits, which occupies 512 MiB, was refused as though it occupied 4 GiB.
One predicate applies the limit after the ceiling division at every tag site,
and the emitted guard measures a bit-size expression the same way.

The wrapper's stored field index was a byte and the container and dynamic
field indices were 16-bit, so a wrapped value behind 256 excluded fields was
read from field zero by the reflection engine while generated code used the
named field. All three are 32-bit.

+++ codegen/parser.go
@@ -1632,7 +1632,7 @@ func (p *Parser) buildContainerDescriptor(desc *ssztypes.TypeDescriptor, dataStr
 		fieldDesc := ssztypes.FieldDescriptor{
 			Name:       schemaField.Name(),
 			Type:       typeDesc,
-			FieldIndex: uint16(runtimeFieldIndex),
+			FieldIndex: uint32(runtimeFieldIndex),
 		}
 
 		// Handle ssz-index for progressive containers - extract from original tag parsing
@@ -1658,7 +1658,7 @@ func (p *Parser) buildContainerDescriptor(desc *ssztypes.TypeDescriptor, dataStr
 			dynFieldDesc := ssztypes.DynFieldDescriptor{
 				Field:        &fieldDesc,
 				HeaderOffset: size,
-				Index:        int16(runtimeFieldIndex), // Runtime field index for data access
+				Index:        int32(runtimeFieldIndex), // Runtime field index for data access
 			}
 			dynFields = append(dynFields, dynFieldDesc)
 			isDynamic = true
@@ -2354,7 +2354,7 @@ func (p *Parser) buildTypeWrapperDescriptor(desc *ssztypes.TypeDescriptor, dataN
 
 	// Store wrapper information
 	desc.ElemDesc = wrappedDesc
-	desc.WrapperFieldIndex = uint8(wrappedFieldIndex)
+	desc.WrapperFieldIndex = uint32(wrappedFieldIndex)
 
 	// The TypeWrapper inherits properties from the wrapped type
 	desc.Size = wrappedDesc.Size
